feat: Add NetBox superuser creation task
This task is not idempotent, therefore is can run only once after the
database creation task.
Flushing the handler during play execution will not execute any handler
after the remaining tasks. Therefore the flush needs to be called after
all deployment tasks have been executed.
The flush is only needed for the createsuperuser task, because the
database must be already configured. Hence, the createsuperuser task 
needs to be at the end.